backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pinecone-client==3.0.0
openai==1.54.3
httpx==0.27.0
python-dotenv==1.0.0
gunicorn==21.2.0

